Move some functions to new hit.c file
